add snapshot testing for denoise
