+++ crates/denoise/.gitignore
@@ -1,0 +1,3 @@
+*.wav
+!data/inputs/*.wav
+data/outputs/*.wav
