+++ plugins/fs-sync/Cargo.toml
@@ -36,6 +36,7 @@ specta = { workspace = true, features = ["serde_json"] }
 
 glob = "0.3"
 hypr-afconvert = { workspace = true }
+hypr-mp3 = { workspace = true }
 rayon = { workspace = true }
 rodio = { workspace = true, features = ["symphonia-all"] }
